+++ src/Git_Bulk_Updater/Actions.php
@@ -23,7 +23,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Class Actions
  */
 class Actions {
-	use Webhooks;
+	// use Webhooks;
 
 	/**
 	 * Load hooks.
@@ -31,6 +31,7 @@ class Actions {
 	 * @return void
 	 */
 	public function load_hooks() {
+		$this->load_js();
 		add_action( 'admin_menu', [ $this, 'add_plugin_menu' ] );
 		add_action( 'admin_init', [ new Updater(), 'update' ] );
 	}
@@ -61,22 +62,41 @@ class Actions {
 	 * Options page callback.
 	 */
 	public function create_admin_page() {
-		$action = is_multisite() ? 'edit.php?action=git-bulk-updater' : 'options.php';
-		?>
-		<div class="wrap">
-			<h2>
-				<?php esc_html_e( 'Git Bulk Updater', 'git-bulk-updater' ); ?>
-			</h2>
-			<form method='post' action="<?php esc_attr_e( $action ); ?>">
-			<table class='form-table'>
-			<tbody>
-		<?php
-		( new Actions_Row() )->add_site_row();
+			$action = is_multisite() ? 'edit.php?action=git-bulk-updater' : 'options.php';
+
+		echo '<div class="wrap"><h2>';
+		esc_html_e( 'Git Bulk Updater', 'git-bulk-updater' );
+		echo '</h2>';
+		$this->repo_or_site_selector();
+		echo '<form method="post" action="' . esc_attr( $action ) . '">';
+		echo '<table class="form-table">';
+		echo '<tbody class="git-bulk-updater-repo">';
 		( new Actions_Row() )->add_repo_row();
-		echo '</tbody></table></div>';
+		echo '</tbody>';
+		echo '<tbody class="git-bulk-updater-site">';
+		( new Actions_Row() )->add_site_row();
+		echo '</tbody>';
+		echo '</table></div>';
 		echo '</form>';
 	}
 
+	/**
+	 * Repo or Site option.
+	 */
+	public function repo_or_site_selector() {
+		$options = ['git-bulk-updater-repo' => 'Show Repos','git-bulk-updater-site'=>'Show Sites'];
+		?>
+		<label for="git-bulk-updater">
+			<select id="git-bulk-updater" name="git-bulk-updater">
+				<?php foreach ( $options as $key => $value ) : ?>
+						<option value="<?php esc_attr_e( $key ); ?>" <?php selected( $key ); ?> >
+							<?php esc_html_e( $value ); ?>
+						</option>
+				<?php endforeach ?>
+			</select>
+		</label>
+		<?php
+	}
 	/**
 	 * Redirect to where we came from.
 	 *
@@ -91,4 +111,20 @@ class Actions {
 		wp_safe_redirect( $location );
 		exit();
 	}
+
+		/**
+	 * Load javascript for Install.
+	 *
+	 * @return void
+	 */
+	public function load_js() {
+		add_action(
+			'admin_enqueue_scripts',
+			function () {
+				wp_register_script( 'git-bulk-updater-actions', plugins_url( basename( GIT_BULK_UPDATER_DIR ) . '/js/git-bulk-updater-switcher.js' ), [], false, true );
+				wp_enqueue_script( 'git-bulk-updater-actions' );
+			}
+		);
+	}
+
 }