+++ CHANGES.md
@@ -3,3 +3,4 @@
 * add rows for both update entire site or update all sites with specific repository
 * add ability to use multiple JSON files.
 * update for use with downloadable JSON files from GitHub Updater
+* add javascript to switch display from repos/sites